Исправлен запуск. Добавляется новое... (#16)

+++ app/src/main/java/com/nightstalker/artic/features/artwork/presentation/ui/list/ArtworksListAdapter.kt
@@ -37,7 +37,7 @@ class ArtworksListAdapter(
             Glide.with(context).load(imageUrl).into(placeImage)
 
             root.setOnClickListener {
-                onItemClicked(item.id.toInt())
+                onItemClicked(item.id)
             }
         }
     }
